requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-groq
pypdf
chromadb
sentence-transformers
streamlit
python-dotenv
pytest
requests
beautifulsoup4
lxml
pytesseract
pdf2image
tqdm
trafilatura
langchain-experimental
rank_bm25
flashrank
